fix(web): fix contacts Enable button navigation losing prompt param (LUM-1696)
The onStartSetupConversation handler navigated to /assistant?prompt=...
which triggered bootstrap routing that replaced the URL with
/assistant/conversations/{key} — stripping the ?prompt= search param.
The prompt was lost, so no message was ever sent to the new thread.

Fix: create a draft conversation ID, set it active in the store, and
navigate directly to the conversation URL with the prompt preserved.
This follows the same proven pattern used by home-page-route.tsx.

Also fixes the BACK arrow by ensuring proper history stack — the push
navigation from /assistant/contacts to the conversation URL means
navigate(-1) correctly returns to contacts.

Extracted ContactsPage into a route-level wrapper (contacts-page-route.tsx)
to satisfy the no-cross-domain-imports lint rule.
